+++ console/cypress/helpers/dataHelpers.ts
@@ -146,6 +146,39 @@ export const testCustomFunctionSQLWithSessArg = (
   };
 };
 
+export const createUntrackedFunctionSQL = (
+  fnName: string,
+  tableName: string
+) => {
+  return {
+    type: 'run_sql',
+    args: {
+      sql: `
+          CREATE OR REPLACE FUNCTION ${fnName}(table_row "${tableName}")
+           RETURNS int
+           LANGUAGE sql
+           STABLE
+          AS $function$
+            SELECT table_row.id
+          $function$
+          `,
+      cascade: false,
+    },
+  };
+};
+
+export const dropUntrackedFunctionSQL = (fnName: string) => {
+  return {
+    type: 'run_sql',
+    args: {
+      sql: `
+          DROP FUNCTION public.${fnName};
+          `,
+      cascade: false,
+    },
+  };
+};
+
 export const getTrackFnPayload = (name = 'customfunctionwithsessionarg') => ({
   type: 'pg_track_function',
   args: {