+++ console/cypress/integration/data/modify/spec.ts
@@ -4,6 +4,8 @@ import {
   getTableName,
   getColName,
   getElementFromAlias,
+  createUntrackedFunctionSQL,
+  dropUntrackedFunctionSQL,
 } from '../../../helpers/dataHelpers';
 
 import {
@@ -11,10 +13,36 @@ import {
   validateCT,
   validateColumn,
   ResultType,
+  dataRequest,
 } from '../../validators/validators';
 import { setPromptValue } from '../../../helpers/common';
 
 const testName = 'mod';
+
+export const passMTFunctionList = () => {
+  const tableName = getTableName(0, testName);
+  dataRequest(
+    createUntrackedFunctionSQL(`${tableName}_id_fn`, tableName),
+    ResultType.SUCCESS
+  );
+  cy.wait(5000);
+  cy.get(getElementFromAlias('modify-table-edit-computed-field-0')).click();
+
+  cy.get(getElementFromAlias('functions-dropdown')).click();
+
+  cy.get('[data-test^="data_test_column_type_value_"]').should(
+    'have.length',
+    1
+  );
+
+  cy.get('[data-test^="data_test_column_type_value_"]')
+    .first()
+    .should('have.text', `${getTableName(0, testName)}_id_fn`.toLowerCase());
+  dataRequest(
+    dropUntrackedFunctionSQL(`${tableName}_id_fn`),
+    ResultType.SUCCESS
+  );
+};
 
 export const passMTCreateTable = () => {
   cy.get(getElementFromAlias('data-create-table')).click();
@@ -30,7 +58,10 @@ export const passMTCreateTable = () => {
   cy.wait(7000);
   cy.url().should(
     'eq',
-    `${baseUrl}/data/default/schema/public/tables/${getTableName(0, testName)}/modify`
+    `${baseUrl}/data/default/schema/public/tables/${getTableName(
+      0,
+      testName
+    )}/modify`
   );
   validateCT(getTableName(0, testName), ResultType.SUCCESS);
 };
@@ -41,7 +72,10 @@ export const passMTCheckRoute = () => {
   // Match the URL
   cy.url().should(
     'eq',
-    `${baseUrl}/data/default/schema/public/tables/${getTableName(0, testName)}/modify`
+    `${baseUrl}/data/default/schema/public/tables/${getTableName(
+      0,
+      testName
+    )}/modify`
   );
 };
 
@@ -92,7 +126,10 @@ export const passMTMoveToTable = () => {
   cy.get(getElementFromAlias(getTableName(0, testName))).click();
   cy.url().should(
     'eq',
-    `${baseUrl}/data/default/schema/public/tables/${getTableName(0, testName)}/browse`
+    `${baseUrl}/data/default/schema/public/tables/${getTableName(
+      0,
+      testName
+    )}/browse`
   );
 };
 
@@ -101,7 +138,10 @@ export const failMTWithoutColName = () => {
   cy.get(getElementFromAlias('modify-table-add-new-column-save')).click();
   cy.url().should(
     'eq',
-    `${baseUrl}/data/default/schema/public/tables/${getTableName(0, testName)}/modify`
+    `${baseUrl}/data/default/schema/public/tables/${getTableName(
+      0,
+      testName
+    )}/modify`
   );
 
   validateColumn(
@@ -116,7 +156,10 @@ export const failMTWithoutColType = () => {
   cy.get(getElementFromAlias('modify-table-add-new-column-save')).click();
   cy.url().should(
     'eq',
-    `${baseUrl}/data/default/schema/public/tables/${getTableName(0, testName)}/modify`
+    `${baseUrl}/data/default/schema/public/tables/${getTableName(
+      0,
+      testName
+    )}/modify`
   );
   validateColumn(
     getTableName(0, testName),
@@ -137,7 +180,10 @@ export const Addcolumnnullable = () => {
   cy.wait(2500);
   cy.url().should(
     'eq',
-    `${baseUrl}/data/default/schema/public/tables/${getTableName(0, testName)}/modify`
+    `${baseUrl}/data/default/schema/public/tables/${getTableName(
+      0,
+      testName
+    )}/modify`
   );
   validateColumn(
     getTableName(0, testName),
@@ -266,7 +312,10 @@ export const passMTDeleteCol = () => {
   cy.wait(5000);
   cy.url().should(
     'eq',
-    `${baseUrl}/data/default/schema/public/tables/${getTableName(0, testName)}/modify`
+    `${baseUrl}/data/default/schema/public/tables/${getTableName(
+      0,
+      testName
+    )}/modify`
   );
   validateColumn(
     getTableName(0, testName),
@@ -281,7 +330,10 @@ export const passMTDeleteTableCancel = () => {
   cy.window().its('prompt').should('be.called');
   cy.url().should(
     'eq',
-    `${baseUrl}/data/default/schema/public/tables/${getTableName(0, testName)}/modify`
+    `${baseUrl}/data/default/schema/public/tables/${getTableName(
+      0,
+      testName
+    )}/modify`
   );
 
   validateCT(getTableName(0, testName), ResultType.SUCCESS);
@@ -292,8 +344,7 @@ export const passMTDeleteTable = () => {
   cy.get(getElementFromAlias('delete-table')).click();
   cy.window().its('prompt').should('be.called');
   cy.wait(5000);
-  // FIXME: change this later
-  // cy.url().should('eq', `${baseUrl}/data/default/schema/public`);
+  cy.url().should('eq', `${baseUrl}/data/default/schema/public`);
   validateCT(getTableName(0, testName), ResultType.FAILURE);
 };